removed unused variable & fixed clippy error
