+++ src/algorithms/mod.rs
@@ -1,0 +1,5 @@
+pub mod bnb;
+pub mod fifo;
+pub mod knapsack;
+pub mod lowestlarger;
+pub mod srd;
