--- src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    println!("let there be coin selection");
-}
